Show status after synchronous purge if it is an error status.
I had a misconfiguration, where all purge requests were denied by
varnish because my ip address was not in the allowed list.  This gave
a status 504 on all purge requests, but this was nowhere visible.  The
controlpanel only showed 'The following items were purged:' with the
list I expected, no errors.