Set browser.handleErrors to False in one test.
This is when requesting an image scale.
On the plone.namedfile branch for plip 1483 I am getting a KeyError 'x-cache-rule'
when checking the browser headers,
where it would be handier to get an Unauthorized because the user is not allowed to see this image.
(The user probably *should* be allowed to see this image,
but that is something to fix in that plip branch.)