Fixed normalizing etags in tests.
The previous code blindly removed the last etag, assuming it was the time based resourceRegistries ETag.
This is true again, after reintroducing the modernized version of this ETag, but it was not true before.

Moved the two versions of the _normalize_etag function to normalize_etag in test_utils.py.
Fixed it there to only remove the last component when the current year is in it.