Define etag for layout property.
For the moment this only defines the etag, without actually using it in a caching profile.
We can do this later.

When the layout changes, the modification date stays the same (currently).
So without this etag, you can get a page from the cache with the old layout.

Initially I had an etag for the default page as well.  This could be added to containers.
But the published item is then not the container, but the page.
And the page itself has no default page, so this etag never returned anything.