Split default/registry.xml into various files.
The general settings remain in registry.xml.
Weak, moderate, strong, terse, and generic get their own file in the registry directory.
(Generic is not used by default, it was commented out in the original.)
This should give a better overview.

Also, this makes it easier to copy `registry/terse.xml` for an upgrade step.