Apply weak caching to GET requests of content with application/json.
See https://github.com/plone/plone.rest/issues/73.

Note: this needs an upgrade step, but we will do this in plone.app.upgrade.
The plone.app.caching master branch already has an update to metadata version 2, so there is no room for an upgrade step on the 2.x branch.

Also note that it should not be needed to undo this change on Plone 6:
there the plone.content.dynamic ruleset will be applied, and this takes precedence over the current change.